services/data-ingestion/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
uvloop==0.19.0

# Data Processing
kafka-python==2.0.2
aioredis==2.0.1
psycopg2-binary==2.9.7
asyncpg==0.29.0

# HTTP Clients
aiohttp==3.9.1
httpx==0.25.0
requests==2.31.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Monitoring and Metrics
prometheus-client==0.19.0
psutil==5.9.6

# Data Validation
pydantic==2.5.0
jsonschema==4.20.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
structlog==23.2.0

# Circuit Breaker
tenacity==8.2.3

# Testing (for development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
